Specify const type.

+++ b30/lib.rs
@@ -10,7 +10,7 @@ use worker::{console_log, event, Context, Env, Fetch, Headers, Method, Request, 
 use worker_kv::KvStore;
 
 const USER_AGENT: &str = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36";
-const CONCURRENT_REQUESTS: usize = 5;
+const CONCURRENT_REQUESTS: u32 = 5;
 const BASE_TAPHUNTER_URL: &str = "http://www.taphunter.com/bigscreen";
 const BASE_UNTAPPD_URL: &str = "https://untappd.com";
 const CACHE_TTL_SECONDS: u64 = 7 * 24 * 60 * 60; // 1 week
@@ -236,7 +236,7 @@ async fn fetch_untappd_ratings(
 
             (idx, rating)
         })
-        .buffer_unordered(CONCURRENT_REQUESTS)
+        .buffer_unordered(CONCURRENT_REQUESTS as usize)
         .collect()
         .await;
